requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpathlib==0.23.0
confection==1.3.3
cryptography==46.0.6
cymem==2.0.13
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.135.2
Flask==3.1.3
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
numpy==2.4.3
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
preshed==3.0.13
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.6.0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
requests==2.33.0
rich==14.3.3
setuptools==82.0.1
shellingham==1.5.4
smart_open==7.5.1
spacy==3.8.13
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
starlette==1.0.0
thinc==8.3.13
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
wasabi==1.1.3
weasel==1.0.0
Werkzeug==3.1.7
wrapt==2.1.2
